Rename 'xml' parser package to 'xpath' in preparation of extending it.
