+++ plugins/parsers/xpath/parser.go
@@ -1,4 +1,4 @@
-package xml
+package xpath
 
 import (
 	"fmt"
