+++ plugins/parsers/xpath/testcases/earthquakes.conf
