+++ plugins/parsers/xpath/testcases/multisensor_explicit_basic.conf
