+++ plugins/parsers/xpath/testcases/multisensor_explicit_batch.conf
