+++ plugins/parsers/xpath/testcases/multisensor_selection_batch.conf
